v2.3 Migration to Staging

+++ src/entities/sampling-train.entity.ts
@@ -27,7 +27,7 @@ export class SamplingTrain extends BaseEntity {
     name: 'rpt_period_id',
     nullable: false,
     transformer: new NumericColumnTransformer(),
-    type: 'numeric',
+    type: 'bigint',
   })
   reportingPeriodId: number;
